move dq_evaluation to utils/evaluation.py
