+++ predict.py
@@ -10,7 +10,7 @@ from keras.utils import CustomObjectScope
 from data_engine.prepare_data import build_dataset, update_dataset_from_file, keep_n_captions
 from utils.callbacks import *
 from quest import models as modFactory
-from quest import dq_evaluation
+from utils import evaluation
 
 logging.basicConfig(level=logging.DEBUG,
                     format='[%(asctime)s] %(message)s', datefmt='%d/%m/%Y %H:%M:%S')
